feat: extract tool classification into classify.rs (#633)
Move pure classification functions (map_tool, operation_to_node_kind,
classify_tool_output, truncate_subject, node_kind_to_u8/u8_to_node_kind,
LeafTracker) and their tests out of main.rs into a dedicated module.
Reduces main.rs by ~500 lines (3845 -> 3339) toward the 1000-line target.
